# openapi.yml
openapi: 3.1.0
info:
  title: JoReminder API
  description: Apple Reminders 웹 클론 REST API
  version: 0.1.0

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /api/lists:
    get:
      summary: 전체 목록 조회
      operationId: findAllLists
      tags: [Lists]
      responses:
        "200":
          description: 목록 배열 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderListResponse"

    post:
      summary: 목록 생성
      operationId: createList
      tags: [Lists]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderListRequest"
      responses:
        "201":
          description: 생성된 목록 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"

  /api/lists/{id}:
    put:
      summary: 목록 수정
      operationId: updateList
      tags: [Lists]
      parameters:
        - $ref: "#/components/parameters/ListId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderListRequest"
      responses:
        "200":
          description: 수정된 목록 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: 목록 삭제
      description: 기본 목록("미리 알림")은 삭제할 수 없다.
      operationId: deleteList
      tags: [Lists]
      parameters:
        - $ref: "#/components/parameters/ListId"
      responses:
        "204":
          description: 삭제 완료
        "400":
          description: 기본 목록 삭제 시도
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/lists/{listId}/reminders:
    get:
      summary: 특정 목록의 리마인더 조회
      operationId: findRemindersByList
      tags: [Reminders]
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 리마인더 배열 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/reminders:
    post:
      summary: 리마인더 생성
      operationId: createReminder
      tags: [Reminders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderRequest"
      responses:
        "201":
          description: 생성된 리마인더 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"

  /api/reminders/{id}:
    put:
      summary: 리마인더 수정
      operationId: updateReminder
      tags: [Reminders]
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderRequest"
      responses:
        "200":
          description: 수정된 리마인더 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: 리마인더 삭제
      operationId: deleteReminder
      tags: [Reminders]
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      responses:
        "204":
          description: 삭제 완료
        "404":
          $ref: "#/components/responses/NotFound"

  /api/reminders/{id}/toggle:
    patch:
      summary: 리마인더 완료 토글
      operationId: toggleReminder
      tags: [Reminders]
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      responses:
        "200":
          description: 토글된 리마인더 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/reminders/today:
    get:
      summary: 오늘 마감 리마인더
      operationId: findTodayReminders
      tags: [Smart Lists]
      responses:
        "200":
          description: 리마인더 배열 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderResponse"

  /api/reminders/scheduled:
    get:
      summary: 마감일 있는 미완료 리마인더 (날짜순)
      operationId: findScheduledReminders
      tags: [Smart Lists]
      responses:
        "200":
          description: 리마인더 배열 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderResponse"

  /api/reminders/all:
    get:
      summary: 전체 미완료 리마인더
      operationId: findAllReminders
      tags: [Smart Lists]
      responses:
        "200":
          description: 리마인더 배열 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderResponse"

  /api/reminders/completed:
    get:
      summary: 완료된 리마인더
      operationId: findCompletedReminders
      tags: [Smart Lists]
      responses:
        "200":
          description: 리마인더 배열 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderResponse"

components:
  parameters:
    ListId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

    ReminderId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    ReminderListRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "장보기"
        color:
          type: string
          example: "GREEN"

    ReminderListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        reminderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ReminderRequest:
      type: object
      required: [title, listId]
      properties:
        title:
          type: string
          example: "우유 사기"
        listId:
          type: integer
          format: int64

    ReminderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        completed:
          type: boolean
        listId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProblemDetail:
      type: object
      description: RFC 9457 Problem Detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string

  responses:
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"